requirements.txt
# 数据处理与分析
pandas>=1.5.0
numpy>=1.22.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 自然语言处理
nltk>=3.7.0
jieba>=0.42.0
transformers>=4.15.0
torch>=1.10.0
langdetect>=1.0.9

# 文档处理
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.10.0
requests>=2.27.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.8.0
plotly>=5.5.0
kaleido>=0.2.0
streamlit>=1.30.0

# 报告生成
jinja2>=3.0.0
markdown>=3.3.0
weasyprint>=54.0

# 其他工具
tqdm>=4.62.0
pyldavis>=3.3.0
textblob>=0.17.0
rich>=12.0.0 